// mini-xrm-frontend/mini-xrm-client-angular/model/activityView.ts
/**
 * Mini xRM API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ActivityView { 
    id: number;
    subject: string;
    type: string;
    description: string;
    durationMinutes: number;
    personResponsible: string;
    partnerId: number;
}